Slit TimedLoop class into file separate from session.py
